refactor(core,utils): 重构临时文件清理逻辑+迁移load_env
